requirements.txt
ailment==9.0.5405
angr==9.0.5405
archinfo==9.0.5405
bcrypt==3.2.0
beautifulsoup4==4.9.1
bitstring==3.1.7
cachetools==4.2.1
capstone==4.0.2
certifi==2020.4.5.1
cffi==1.14.4
chardet==3.0.4
claripy==9.0.5405
cle==9.0.5405
CppHeaderParser==2.7.4
cryptography==3.3.1
DateTime==4.3
decorator==4.4.2
dpkt==1.9.4
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
idna==2.9
intervaltree==3.1.0
itanium-demangler==1.0
Mako==1.1.4
MarkupSafe==1.1.1
mulpyplexer==0.9
networkx==2.5
numpy==1.18.4
packaging==20.8
paramiko==2.7.2
pefile==2019.4.18
pexpect==4.8.0
plumbum==1.6.9
ply==3.11
progressbar2==3.53.1
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
pyelftools==0.27
Pygments==2.7.4
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PySMT==0.9.0
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.5.5
pytz==2020.1
pyvex==9.0.5405
requests==2.23.0
ROPGadget==6.5
rpyc==5.0.1
six==1.15.0
smmap==3.0.5
sortedcontainers==2.3.0
soupsieve==2.0.1
unicorn==1.0.2rc4
urllib3==1.25.9
wllvm==1.2.8
z3-solver==4.8.10.0
zope.interface==5.1.0
pwntools==4.10.0
